removed need to use injections
