+++ play/Play.vue
@@ -1,7 +1,6 @@
 <template>
     <div>
-      <iv-title-bar>The Playground</iv-title-bar>
-      <header class="appheader" style="padding-top: 3vh;">
+      <header class="appheader">
           <div class="start-items">
             <h1>Playground!</h1>
             <div class="app__components-select">
@@ -11,8 +10,6 @@
             </div>
           </div> 
       </header>
-      
-
       <sandbox class="app__content">
           <component :is="getComponent"></component>
