+++ src/components/index.js
@@ -6,7 +6,6 @@ export {SVGContainer, SVGDraggable} from "./SVGDragAndDrop";
 export {default as EquationBox} from "./EquationBox";
 export {default as EquationBoxAdvance} from "./EquationBoxAdvance";
 export {default as HoverText} from "./HoverText";
-export {default as TitleBar} from "./TitleBar";
 export {default as DropdownTextBox} from "./DropdownTextBox";
 export {default as DropdownList} from "./DropdownList";
 export {default as ToggleBasic} from "./ToggleBasic";
