+++ src/layouts/TitleBar/index.js
@@ -1,0 +1,2 @@
+import TitleBar from "./TitleBar.vue";
+export default TitleBar;
