+++ src/layouts/Visualisation/Visualisation.vue
@@ -7,8 +7,8 @@
 //Potentially at some point this should be renamed to something like visulaisation manager? Or something to make it explicit that this object manages the state of the panels and dropouts etc...
 export default {
     name:"iv-visualisation",
-    provides(){
-        return {
+    data(){
+        return{
             reservedSlots:[]
         }
     }
