+++ src/layouts/index.js
@@ -1,4 +1,5 @@
 export {default as MainStage} from './MainStage';
 export {default as Visualisation} from './Visualisation';
 export {default as Pane} from './Pane';
+export {default as TitleBar} from './TitleBar';
 export {default as FixedBlob} from './Hotspotables';
